+++ src/WebJobs.Extensions.CosmosDB/release_notes.md
@@ -1,6 +1,5 @@
-## Microsoft.Azure.WebJobs.Extensions.CosmosDB 4.16.0
+## Microsoft.Azure.WebJobs.Extensions.CosmosDB 4.16.1
 
-- feat: add support for `AllVersionsAndDelete` changefeed mode. (#999)
-- feat: update Microsoft.Azure.Cosmos to 3.60.0 (#999)
+- fix: wrap worker-default trigger type as ChangeFeedItem<JObject> for AllVersionsAndDeletes (#1000)
 
 **NOTE**: The stable version of this package does not include CosmosDB preview features. Use the `-preview` versions of this package for CosmosDB preview service features.